+++ src/compas_ghpython/components/Compas_FromJson/code.py
@@ -8,6 +8,9 @@ import compas
 
 class CompasInfo(component):
     def RunScript(self, json):
+        if not json:
+            return None
+
         try:
             return compas.json_load(json)
         except:    # noqa: E722